% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatFunctTest.R
\name{CatFunctTest}
\alias{CatFunctTest}
\title{Analysis of Categorical Variables}
\usage{
CatFunctTest(dat, names)
}
\arguments{
\item{dat}{A data frame}

\item{names}{A vector of names of your categorical variables}
}
\value{
A data frame with n, percentage, and p-value (from other Fisher's or Chi-Square tests)
for each variable and group
}
\description{
Analysis of Categorical Variables
}
