% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContFuncTest.R
\name{ContFuncTest}
\alias{ContFuncTest}
\title{Analyze Continuous Variables}
\usage{
ContFuncTest(dat, names)
}
\arguments{
\item{dat}{A data frame}

\item{names}{A vector of names of continuous variables}
}
\value{
A data fram with mean, standard deviation, and p-value (either t-Test or U-test) for each
group and variable
}
\description{
Analyze Continuous Variables
}
