% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertToCat.R
\name{ConvertToCat}
\alias{ConvertToCat}
\title{Converting Categorical Variables to Factors}
\usage{
ConvertToCat(df, nam)
}
\arguments{
\item{df}{A data frame}

\item{nam}{A vector of names of your categorical variables}
}
\value{
Variables as factors
}
\description{
Converting Categorical Variables to Factors
}
