% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myCatTable.R
\name{myCatTable}
\alias{myCatTable}
\title{Generate the Categorical Variable Table}
\usage{
myCatTable(dat)
}
\arguments{
\item{dat}{Your output from CatFunctTest}
}
\value{
A table with your analysis
}
\description{
Generate the Categorical Variable Table
}
