% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myContTable.R
\name{myContTable}
\alias{myContTable}
\title{Generate the Continuous Variable Table}
\usage{
myContTable(dat)
}
\arguments{
\item{dat}{Your output from ContFuncTest}
}
\value{
A Table with your analysis
}
\description{
Generate the Continuous Variable Table
}
